feat: dj-events-nextjs project - implementing the edit component

+++ dj-events-frontend/pages/events/[slug].js
@@ -6,9 +6,11 @@ import { useRouter } from 'next/router'
 import superjson from 'superjson'
 
 import MyLayout from '@/components/MyLayout'
-import { API_URL } from '../../config'
+import { API_URL } from '@/config/index'
 import styles from '@/styles/Event.module.css'
-import { dateToLocalString } from '../../utils/utils_main'
+import { dateToLocalString } from '@/utils/utils_main'
+import { useState } from 'react'
+import { ToastHelper } from '@/utils/toast_helper'
 
 export async function getStaticProps({ params }) {
   const res = await fetch(`${API_URL}/events?slug=${params.slug}`)
@@ -47,21 +49,43 @@ export async function getStaticPaths() {
 }
 
 const EventPage = ({ myEvent }) => {
+  const [toast, setToast] = useState({
+    message: 'Click any actions buttons or just sit tight',
+    type: 'info',
+  })
   const router = useRouter()
   const { slug } = router.query
 
-  const deleteEvent = async () => {
-    const res = await fetch(`${API_URL}/api/events/${slug}`, {
-      method: 'DELETE',
-    })
-    const myEventFromApi = await res.json()
+  const deleteEvent = async (deletedEvtId) => {
+    if (window.confirm('Are you sure you want to delete this event?')) {
+      const res = await fetch(`${API_URL}/events/${deletedEvtId}`, {
+        method: 'DELETE',
+      })
 
-    console.log('deleteEvent myEventFromApi:', myEventFromApi)
-    router.push('/events')
+      const myEventFromApi = await res.json()
+      console.log('[slug] page, deleteEvent data:', myEventFromApi)
+
+      if (!res.ok) {
+        setToast({
+          message: 'Error deleting event. Details:' + myEventFromApi.message,
+          type: 'error',
+        })
+        return
+      }
+
+      setToast({
+        message: 'Event deleted successfully',
+        type: 'success',
+      })
+
+      router.push('/events')
+    }
   }
 
   return (
     <MyLayout>
+      <ToastHelper message={toast.message} type={toast.type} />
+
       <div className={styles.event}>
         <div className={styles.controls}>
           <Link href={`/events/edit/${myEvent.id}`}>
@@ -69,7 +93,11 @@ const EventPage = ({ myEvent }) => {
               <FaPencilAlt /> Edit Event
             </a>
           </Link>
-          <a href="#" className={styles.delete} onClick={deleteEvent}>
+          <a
+            href="#"
+            className={styles.delete}
+            onClick={() => deleteEvent(myEvent.id)}
+          >
             <FaTimes /> Delete Event
           </a>
         </div>
